finops/reverse-etl/requirements.txt
psycopg2-binary>=2.9,<3
simple-salesforce>=1.12,<2
